Finito Printf

+++ PRINTF/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: pcatapan <pcatapan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/27 16:19:34 by pcatapan          #+#    #+#             */
-/*   Updated: 2022/01/28 03:07:32 by pcatapan         ###   ########.fr       */
+/*   Updated: 2022/01/28 08:45:03 by pcatapan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,11 +20,11 @@
 
 int				ft_printf(const char *str, ...);
 int				ft_char(char lett);
-unsigned int	ft_lenunsi(unsigned long num);
-int				ft_unsigned(unsigned long num);
+unsigned int	ft_lenunsi(unsigned int num);
+int				ft_unsigned(unsigned int num);
 unsigned int	ft_esadecimal(unsigned int num, char lett);
 int				ft_putnumber(int num);
-unsigned long	ft_lenesa(uintptr_t nb);
+int				ft_lenesa(uintptr_t nb);
 int				ft_putesa(uintptr_t num);
 
 #endif
